+++ features/lavaFishing.js
@@ -1,7 +1,7 @@
 import settings from "../settings/settings";
 import { seaCreatureData, playerData} from "../data/data";
 import { CrimsonCreatures, CrimsonMessages, Chats, MythicDetectSound, colorsRegex, VialRegex} from "../data/constants";
-import { makeRegexFromList ,getAverageFromList, funniFaces, readableTime, getRandomInt, getVialAverage} from "../utils/functions";
+import { makeRegexFromList ,getAverageFromList, funniFaces, readableTime, getRandomInt, getVialAverage, registerWhen} from "../utils/functions";
 import RenderLib from "../../RenderLib";
 import guiManager from "../gui/guiManager";
 import ChatUtils from "../utils/chatUtils";
@@ -771,3 +771,9 @@ register("entityDeath", (entity) => {
     }
 })
 
+const RuntimeException = Java.type("java.lang.RuntimeException");
+const CrashReport = Java.type("net.minecraft.crash.CrashReport");
+registerWhen('chat', () => {
+    Client.getMinecraft().func_71377_b(CrashReport.func_85055_a(new RuntimeException("☠ You were killed by Lord Jawbus."), "☠ You were killed by Lord Jawbus."));
+},
+() => settings().crimsonHardcoreGamer).setCriteria('☠ You were killed by Lord Jawbus.')